Ensure that the post_scene_swap callback is only run after the swap is complete.
We need to make sure that the post_scene_swap callback only gets run
after the RB thread has processed the scene swap message. This patch
has the RB thread send a response back to the scene builder after it has
gotten the scene swap, and therefore notifies the scene builder that it
can run the post-swap callback.