Improve border dot handling for small radii
When rendering dotted borders with small radii, put a single dot in the
corner and render the edge based on that. This fixes an issue where dots
smoosh into each other with small radii.

The resulting rendering isn't perfectly aligned for the moment, but this
can be improved as we handle more complicated situations such as
combinations of borders of different widths.