Use a freelist to store render task cache entries, make cache entry public.
This is a small patch in preparation for upcoming changes that are
related to caching render tasks. Storing these as a freelist allows
handing out weak handles, rather than always fetching by hash key.

We can provide these to the caller of request_render_task() to allow
fetching the CacheItem during batching, without another hash map
lookup. This is more in line with how the PathFinder integration
works, and also useful for some upcoming patches which will cache
other render task kinds. Specifically, in addition to the CacheItem
for a cached render task, the batching code will be able to retrieve
other details about a cached render task, such as required blend mode.