Add API on stacking contexts to configure glyph raster mode.
This allows configuring per-stacking context whether glyphs are
rasterized in screen space or local space.

Also add a reftest for the raster modes.

In the future, we may want to change this to be a more general
raster mode for all primitives and/or allow an individual scale
value for x/y in local space mode.