[Optimization] Collapse opacity filters where possible into primitive.
In many cases, an opacity filter is used to make a rectangle or
image partially transparent, and only contains a single primitive.

In these cases, we can skip drawing the primitive to an off-screen
surface and applying the opacity while blitting that target to the
parent surface.

Instead, we collapse the opacity value(s) into the color of the
primitive, and skip the intermediate surface altogether, drawing
the primitive directly into the parent surface.

This is a very common use case on real world sites, so it has a
significant performance win in many sites.