Robustify shutdown for the scene builder thread
Currently the scene builder thread can be processing a scene when the RB
thread shuts down. In this case, the scene builder thread may end up
trying to unwrap() the send result from the
SceneBuilderResult::Transaction, but that would fail and panic. Although one
alternative is to simply remove that unwrap() call, it seems better to
have a proper shutdown sequence and assert that things are happening as
we expect them to.