Only run the pre- and post-swap hooks for transactions that build a scene.
All sorts of transactions go through the scene builder thread, and only
some of them actually build a scene and result in a scene swap. It's
only in those cases that we need to invoke the pre- and post-swap hooks.
Doing it in other cases is wasteful, for Gecko at least.