Support non-square slab allocator pages in texture cache.
Text runs are typically rectangles that are much longer in one
dimension than the other dimension (depending on horizontal or
vertical layout direction).

Storing rectangles of this shape in the texture cache was very
inefficient, since only square allocations are supported.

This patch adds support for texture cache regions to support
non-square slab sizes.

This is an efficiency win for any image that is rectangular,
and will become particularly important once we start caching
text runs in the texture cache.