Prevent sampling concurrently with the post-swap hook running.
If we run the sampler hook on the render backend thread concurrently
with the post-swap hook running on the scene builder thread, we can get
inconsistent sampling results. Refer to
https://bugzilla.mozilla.org/show_bug.cgi?id=1457470 for more details;
this is a Gecko-specific problem because the post-swap hook is used to
update data structures from which the sampling hook reads.

This patch adds an extra channel that serves as a barrier so that during
the scene swap, we wait until the post-swap hook is done before
resuming the render backend to the point where it calls the sampler
hook. This effectively enforces that the scene builder and render
backend perform the scene swap operation in lockstep which is what we
want. Although not great from a WebRender portability point of view, all
this only happens if the swap hook is installed, so it shouldn't affect
other users of WebRender.