Store stretch size in primitives, free up extra segment block.
The data that was being stored in the segment extra data field is
constant, per segment. Instead, store the stretch size in each
primitive as appropriate, and derive the repeat count in the vertex
shader from the local rect size and stretch size.

This makes a complete 4 floats available in the per-segment data,
which I intend to use as part of porting borders to be brushes.