Add initial brush border primitive.
This is currently only used for image borders. We can incrementally
port more border types to use this brush primitive, and then we
can eventually remove the old border shader and primitive type.

This is also a small optimization for image borders. Previously
image borders created 9 top level primitives via add_image, which
involves extra culling, clip-chain tests etc.

This also removes the only use of sub_rect in the add_image API.
We can probably remove support for that as a follow up. I have
left that in place for now due to other in-progress work that
is modifiying the image primitive APIs.