Refactored glyph_rasterizer into folder with seperate modules
This commit doesn't change any semantics, but it exposes the files
`glyph_rasterizer_ext_no_pathfinder` (containing only the items
previously marked with cfg(not(feature = "pathfinder")) and the
`glyph_rasterizer_pathfinder`, which has all previous
cfg(feature = "pathfinder") items.

The public / private visibilities of the structs are the same as before.