Add an explicit API to add reference frames
Previously reference frames were inserted into the ClipScrollTree when a
stacking context had a transformation or a perspective. This changes
makes the API for adding reference frames explicit in order to give
clients better control over when and how they are created. This change
also opens up future opportunities for shrinking the size of the display
list (by separating out the ClipScrollTree from the list of display
items) and for concurrency during the flattening process.