Don't intersect the dirty rect with the tile rect.
Gecko's blob renderer assumes that the dirty rect is in image
space and not in tile space. To avoid having to translate back
to image space in Gecko we can just not clip to the tile in WebRender.

This also fixes the wrench's blob renderer to store the tile size
and clip to the tile rect before rendering.