Remove StackingContext::scroll_policy
This can be handled in clients now that they have access to reference
frame ciip ids. Removing fixed positioning support from WebRender
simplifies the code greatly and also opens the way to explicit creation
of reference frames in the API. It will also allow us to get rid of the
hacky replacements code during display list flattening.

This also requires exposing the root reference frame to Wrench tests in
order to enable them retain the fixed positioning feature. This change
exposes new aliases to wrench "root-scroll-node" and
"root-reference-frame" in order to make this clearer. These nodes
correspond to the ids 1 and 0 respectively and Wrench will halt if it
encounters these used for other nodes.