Properly update the viewport rect on scroll nodes
I introduced a regression in PR #2743 with the handling of updated
scroll frames. Instead of the new viewport size, the old one is used.
This lead to misbehavior of sticky elements in Gecko. Unfortunately, I
don't know a good way to test this with an automated test, but I've
confirmed that this fixes the issue manually.