Ensure epoch updates and callbacks fire for zero-pixel frames
Even if the window size is zero in some dimensions (which happens
legitimately in Gecko), we don't want to bail out completely at the
scene building step. We need to pretend to do the work, and send the
necessary epoch updates and callbacks back to Gecko. If we fail to do
this Gecko can end up in a stuck state where it is waiting for a render
that it thinks is going to happen but never does. This patch optimizes
out the rendering step in these cases, rather than doing an early-exit
on the entire codepath.