Add initial port of borders to brush primitives, with task caching.
This adds the infrastructure for drawing borders as brushes. The
border primitive is drawn via a render task, and cached between
both scroll frames and also new display lists, since the cache key
is based on the content of the border.

The border is drawn as a nine-patch, in order to minimize the
amount of space that is allocated in the texture cache for the
primitive. The edge segments use the repetition support in the
brush shader to draw the complete edge. This makes normal CSS
borders behave almost exactly the same in the rendering phase
as image-borders, which improves batching (since they are drawn
with a simple image shader).

This initial patch supports only a small subset of the possible
border primitives - once this lands I'll be incrementally adding
the other border styles to this system, and eventually removing
the legacy border primitive.

I had to add some fuzziness to a couple of the reftests, due to
the test running through the new path, vs. the reference running
through the old border path, which has slight differences in the
AA between edges. Once all the border types are ported to the new
path, we can remove this fuzziness.