Add a hook for resource-update-only SB transaction completion.
The scene builder might process a transaction which has only resource
updates and no actual display list. In this case we still need to
trigger a composite after the transaction is processed, but we don't
need to the more heavyweight stuff that goes with a scene swap. This
patch adds a new hook that enables the caller to trigger the composite.