Defer building of instances unless render task is stale.
Also reduce clones of the brush segments.

This is an optimization for the common case of a lot of similar
borders. However, it mainly prep work for supporting dashing and
dotting style support. In these cases, we may generate a large
number of instances, and so it would be inefficient to generate
these unconditionally. Instead, it's much more efficient to only
generate them when there is no matching valid cached render task.