Port dot/dash to border brushes, remove old border code.
This ports the last remaining legacy border styles to run with
the new border brush implementation.

With the complete, it's possible to remove all the old border
path code and shaders.

The dash/dot placement is not quite the same as the previous
code. It's better in some cases, and worse in others. Working
on this as a follow up to handle all the possible edge cases
should be much easier once this patch lands.

This completes the work of moving borders to be brush images,
which gives some very significant performance wins - mostly
stemming from (a) caching of border primitives across scrolls
and display lists and (b) much better batching during the main
scene pass, since all borders are now batched as brush_images.