Always do a render after swapping in a new scene.
The render ensures that the hit-tester is rebuilt using the most recent
display list (which is installed into gecko's APZ code at the same time
as the scene swap), along with the corresponding async scroll offsets
and scrollbar thumb transforms (which are sampled from gecko's APZ step
and used to update the clip-scroll tree during a render).

The render step is triggered via a new argument to update_document,
because the existing generate_frame flag on the transaction has semantic
meaning (in that it triggers a callback) that we don't want to interfere
with.

In addition, we can remove the obsolete `render_on_hittest` flag.