Remove TRANSFORM variant of text run shader.
This is a step towards converting the text run shader to be
a brush shader. Like brush shaders, we now determine whether to
apply pixel snapping based on a conditional, rather than a new
shader and batch break.

Remove TransformBatchKind enum, since it's no longer used.

Change the color of B_Blend shader in the profiler, since the
existing color was very close to the color used by the
B_LinearGradient shader, and hard to distinguish between them.