Unify show, load, and replay Wrench commands
The three commands all load different kinds of traces and recordings.
Since we know what behavior we want based on the argument given, we can
unify these three commands.

Currently, the arguments are still specific to each particular recording
type, but eventually they can be made functional for all of them or
removed.