Force the render if and only if a new scene is swapped in.
We need to do a render (or more precisely, rebuild the hit tester) after
an async built scene is swapped in. Prior to this patch, we were
triggering the render on non-empty transactions, but that could include
transactions that didn't actually swap in a scene (i.e just had resource
updates) and could skip cases where we *did* swap in a scene but had an
otherwise empty transaction. The former case led to unnecessary renders
and the latter case led to correctness problems. This patch makes it so
that we trigger the render if and only a new async built scene is
swapped in.