Cache the used font instance during prepare, for batching.
Instead of calling get_font() during batching, this is now
stored as a used_font field in the text run primitive.

This is a small optimization, but is mostly tidy up work in
preparation for some upcoming picture caching patches. The
reason for this change is to ensure that the batching code
doesn't need to access the *value* of the scroll node
matrices, which will be important once pictures can be
rasterized in different spaces.