Fix crashtest panic in ResourceCache::request_image.
ResourceCache::update_texture_cache checks if the request has a tile
associated with it, and if so, adjusts the descriptor and dirty rect for
the individual texture updates. If the image should have been tiled, but
the request did not provide such information, it still attempts to
proceed without tiling.

ResourceCache::request_image was changed to verify that the dirty rect
for the associated image intersects with the request, to avoid
requesting updates for tiled requests which are unchanged. Before it
only did this check for blob images, but it is now necessary for
external images as well which can provide a dirty rect.

We assume here that all requests for an image that must be tiled will have
the necessary tiling information provided, but that does not appear to be
the case. Clipped and YUV images do not provide tiling information as an
example. This patch reverses the check ordering to be consistent with
update_texture_cache.