Copy all the pipeline epoch information to the new scene.
Without this patch, the pipeline_epoch info on the new scene contains
just the pipelines that were "built" as part of the scene build, plus
pipelines on for which update_epoch was called. This is kind of random,
and there doesn't appear to be any real reason to omit some of the
pipelines. For gecko it is desirable to the epoch information for all
pipelines, even the ones that weren't built, and for consistency we can
just always propagate the epoch info over to the new scene during the
scene build.