Reset the active shader to whatever it was before after rendering
Some external toolkits (such as glium) can't see that webrender
has changed the shader under their feet. glium, for example,
assumes that the active shader program hasn't changed, but in
reality, webrender sets the active shader to 0 after a draw call,
so glium::use_program() doesn't do anything because glium thinks
that the original shader is still active.

Generally, to be a good citizen in OpenGL-land, it's not a good
idea to leave and state changes behind after one frame - the
webrenders rendering should be invisible to any other toolkits,
otherwise workarounds around this no-cleanup "feature" would be
required.