Tidy up SpatialNodeIndex and TransformIndex.
TransformIndex is no longer needed, now that the clip-scroll tree
has been split, and we have SpatialNodeIndex implemented.

The basic idea going forward (not quite implemented yet) is:
 - SpatialNodeIndex is a unique id for a positioning node (per DL).
 - From the TransformPalette, code can retrieve a TransformId, by
   supplying a SpatialNodeIndex, and *optionally* a coordinate
   system ID that this transform should be relative to.

The relationship between SpatialNodeIndex and a TransformId is
currently 1:1, but with the changes above, that will be 1:n.

This will allow code to optionally rasterize each Picture using
a different coordinate space, as required.