Add Transform struct, decouple some prim code from SpatialNode.
The transform information in a SpatialNode contains a complete
local -> screen space transform. In the future, we want to be
able to selectively rasterize certain pictures in local space
(e.g. in a nested 3d transform with perspective).

To achieve this, TransformPalette will be extended to allow
construction of a Transform relative to a given coordinate
system / spatial node.

This patch starts decoupling the culling / prim CPU code from
directly referencing a SpatialNode, and instead retrieving the
relevant information from a Transform provided by the
TransformPalette.

Also rename some instances of scroll_node_id to spatial_node_index
as appropriate.