Decouple batching from clip-scroll tree.
A couple small tidy ups in the spatial node code, and also remove
the last dependency during batching on direct references to the
clip-scroll tree.

Once this is complete for culling, we'll be able to selectively
pick which coordinate system an offscreen surface (picture) is
rendered with, via the TransformPalette.