Add a mechanism to rebuild the hit tester without a full render.
This path just updates the clip scroll tree and hit-tester, without
doing all the extra render work needed for a full frame render. When we
do an async scene swap, we can use this instead of doing a full render
to get the hit-test information in sync with Gecko.

Fixes #2807