Add ClipStore struct, and use flat vec for clip sources.
We previously used a free list for clip sources, since it was
possible to add/remove clip sources lists within a single
display list. That's no longer used, so switch to a flat vec
which is a bit simpler and slightly more efficient.

Add a ClipStore struct. This currently holds clip sources only,
but as we start to support clip chains in different coordinate
systems, we'll move the clip chains from the CST to live here.