Optimize how local rects are calculated during culling.
Apply similar logic to the clip code to minimize the
amount of work that is needed when building local
rects for a picture. This change is quite a
significant optimization - in most cases it halves
the amount of work done here, by skipping the extra
rect calculation for non-3d rendering context
pictures. It also saves some rect x matrix transforms
in various cases.

This is a bit more preparation work for rasterizing
pictures in difference coordinate spaces. There are a
few other miscellaneous improvements too:

 * Rename some fields to be more appropriate.
 * Reduce number of matrix inverse calls.
 * Remove reference_frame_stack, which is no longer
   required.
 * Remove BuiltDisplayList from PrimitiveContext, which
   will be helpful in future patches.