Add support for getting relative transforms between coord systems.
This is useful for mapping coordinate systems during primitive
culling - it avoids the use of matrix inverse so it's no
longer fallible.

It's also noticeably more accurate since we're converting
between a minimal set of coordinate systems, rather than trying
to undo a complete world transform (I verified this in both
wrench and various gecko reftests).

Once this lands, we can use it in more places once we start
rasterizing pictures in local space - we can use this to get
the local-space transforms relative to a given coordinate system
which we'll pass to the shaders.