Fix bug where brush mask kind changes during scrolling.
Calculate the brush clip mask kind during clip chain instance
building, instead of during segment building.

If this changes during a scroll of the same display list, detect
that and rerun the logic that assigns the clip mask kind to
each segment.

This doesn't actually re-segment, just updates the clip mask kind,
which is fine since the segments are based only on clips in the
same spatial node (so unable to change relative to each other
during scrolling).

This fixes https://bugzilla.mozilla.org/show_bug.cgi?id=1483568.

Unfortunately we don't have enough scroll testing infrastructure
in wrench to add a regression test for this yet.