Use get_relative_transform for clip <-> prim space conversions.
This expands the functionality in get_relative_transform to work
for both primitive and clip space conversions.

This code path is not super efficient (it could be improved), but
on the majority of sites it never runs, and even on cases where it
does run, the coordinate system tree is typically very shallow. We
can optimize this function if it ever shows up in a profile.