Various improvements towards correct handling of 3d transforms.
* Pictures only calculate a local rect if they are not a pass-through primitive (i.e. if they have a surface).
* Change batching code to use world rects. This avoids the need to do int conversions, and avoids the need to offset each bounding rect.
* Change prim metadata to store world rects in floats, rather than integer screen rects. This integrates better with the plane splitting code.
* Change clipping code to project to world space for complex clips.
* Introduce SpaceMapper, a struct for efficiently converting between arbitrary coordinate systems.
* Introduce Picture coordinate space, which is used to hold the local rect for a given Picture which has a render surface.
* Prim metadata now only stores the clipped world rect. Most of the time the unclipped world rect isn't needed. It's only needed for Picture primitives that are drawing to a surface - so we calculate it on demand in Picture::prepare_for_render().

The major change here is that pictures now only calculate a local rect if they have a surface. Previously, all pictures tried to calculate a local rect. This caused problems because we were trying to map rects with partial transforms, and losing information as we then only considered the 2d projection within each picture. Now, we only calculate a local rect for pictures that actually have a surface. This fixes several reftests.